Point to V20 of the sports example.
Currently the Sport's example is loading v 0.20.3 libraries which has a bug in sorting. I have JSFiddle with the correct version so feel free to use that or create a new one and point to the right version. If you sort the "Current value" column you can see what I mean.